Allow starting in middle of process

+++ py/fullset.sh
@@ -10,28 +10,39 @@ USCCOREVER=101
 USCBRANCH=b$USCCOREVER-$USCNUM-rp-$USCRP1-$USCRP2
 echo !!!!!!! BUILDING $USCBRANCH
 
-pushd ../uscode-software
+if [ "$USCSTEP" = "" ]; then
+  USCSTEP = 01
+fi
 
-echo To run:
-echo python py/process_xml.py --i=/Users/dev1/Downloads/xml_uscAll\@$USCRP1-$USCRP2.zip --rp1=$USCRP1 --rp2=$USCRP2 --o=../uscode/ --notice=NOTICE --title $USCTITLES
-time python py/process_xml.py --i=/Users/dev1/Downloads/xml_uscAll\@$USCRP1-$USCRP2.zip --rp1=$USCRP1 --rp2=$USCRP2 --o=../uscode/ --notice=NOTICE --title $USCTITLES
+if [ "$USCSTEP" = "01" ]; then
+  USCSTEP = 02
+  pushd ../uscode-software
 
-popd
+  echo To run:
+  echo python py/process_xml.py --i=/Users/dev1/Downloads/xml_uscAll\@$USCRP1-$USCRP2.zip --rp1=$USCRP1 --rp2=$USCRP2 --o=../uscode/ --notice=NOTICE --title $USCTITLES
+  time python py/process_xml.py --i=/Users/dev1/Downloads/xml_uscAll\@$USCRP1-$USCRP2.zip --rp1=$USCRP1 --rp2=$USCRP2 --o=../uscode/ --notice=NOTICE --title $USCTITLES
 
-pushd ../uscode
-
-git checkout master
-git reset --hard HEAD
-
-git checkout -b $USCBRANCH
-
-USCTITLES="$USCTITLES" USCFAILEDTITLES="$USCFAILEDTITLES" sh ../uscode-software/py/update-repo.sh
-
-git push --set-upstream origin $USCBRANCH
-
-git tag t-$USCBRANCH
-
-git branch -f master HEAD ; git push --all ; git push --tags
+  popd
+fi
 
 
-popd
+if [ "$USCSTEP" = "02" ]; then
+  pushd ../uscode
+
+  git checkout master
+  git reset --hard HEAD
+
+  git checkout -b $USCBRANCH
+
+  USCTITLES="$USCTITLES" USCFAILEDTITLES="$USCFAILEDTITLES" sh ../uscode-software/py/update-repo.sh
+
+  git push --set-upstream origin $USCBRANCH
+
+  git tag t-$USCBRANCH
+
+  git branch -f master HEAD ; git push --all ; git push --tags
+
+
+  popd
+  USCSTEP = 03
+fi
